Add a collection of various constants
A lot of tests you want to be sure what goes in one end comes out the
other, so these constants give you a quick way to do without manually
managing the values. Helps with mock objects, and when defining one part
of the test in one function, and the verification in another.