split MappedReader into it's own assembly, with nuspec
This is the only bit from the codebase we're really sharing, so used
as a test bed for nuget.

Making the nuspec file wasn't hard, but keeping it up to date is a job
for a build task.

Once all the moving pieces were in place, it's pretty easy. Since
net35 is our lowest target, just left it there, not going to worry
about publishing for multiple .NET frameworks since this code doesn't
change between frameworks.

refs #717 (1.5)